Ajustes en validaciones para las respuesta de los examenes de los cursos.
