+++ src/ActionsData/CoursesActions.ts
@@ -1,4 +1,5 @@
 import _ from 'lodash';
+import { Response } from 'express';
 import { getNamesUsersList } from './UsersActions';
 import {
   ICourseComments, ICourseCommentsObject,
@@ -436,4 +437,122 @@ export async function getCoursesDataUser({ query } : any) : Promise<ICourseUserL
     created_at: course.created_at,
     updated_at: course.updated_at,
   }
+}
+
+// =====================================================================================================================
+
+/*
+  Others functions
+ */
+
+export function returnNotFound(res: Response, code: string | null) : Response {
+  const ret: any = { msg: 'Respuesta no determinada.' };
+  let statusCode = 500;
+
+  if (code === '404Content') {
+    ret.msg = 'Disculpe, pero el contenido seleccionado no existe o ya no se encuentra disponible.';
+    statusCode = 404;
+  }
+  else if (code === '404Course') {
+    ret.msg = 'Disculpe, pero el curso seleccionado no existe o ya no se encuentra disponible.';
+    statusCode = 404;
+  }
+  else if (code === '404Comment') {
+    ret.msg = 'Disculpe, pero el comentario no existe o no se encuentra disponible.';
+    statusCode = 404;
+  }
+  else if (code === '404CourseUser') {
+    ret.msg = 'Disculpe, pero no ha registrado el curso en su listado.';
+    statusCode = 404;
+    ret.addCourse = true;
+  }
+  else if (code === '404GetDataTemaryUser') {
+    ret.msg = 'Disculpe, pero no se logró encontrar la relación de la prueba en su cuenta.';
+    statusCode = 404;
+  }
+  else if (code === '404Theme') {
+    ret.msg = 'Disculpe, pero el tema seleccionado no existe o no se encuentra disponible.';
+    statusCode = 404;
+  }
+  else if (code === 'notFinishTheme') {
+    ret.msg = 'Disculpe, pero no puede realizar la prueba hasta haber completado todo el contenido del tema.';
+    statusCode = 403;
+  }
+  else if (code === 'errorAction') {
+    ret.msg = 'Disculpe, pero no se logró determinar la acción a realizar.';
+    statusCode = 422;
+  }
+  else if (code === 'errorCommentId') {
+    ret.msg = 'Disculpe, pero el comentario seleccionado es incorrecto.';
+    statusCode = 422;
+  }
+  else if (code === 'errorComment') {
+    ret.msg = 'Disculpe, pero el comentario debe cumplir con los siguientes parámetros: 1. Letras o números (az-AZ 0-9) y los siguientes caracteres especiales: .,#*?¿¡!()\\-+"\'/@.';
+    statusCode = 422;
+  }
+  else if (code === 'errorGroupId') {
+    ret.msg = 'Disculpe, pero el grupo seleccionado es incorrecto.';
+    statusCode = 422;
+  }
+  else if (code === 'errorThemeId') {
+    ret.msg = 'Disculpe, pero el tema seleccionado es incorrecto.';
+    statusCode = 422;
+  }
+  else if (code === 'errorContentId') {
+    ret.msg = 'Disculpe, pero el contenido seleccionado es incorrecto.';
+    statusCode = 422;
+  }
+  else if (code === 'like') {
+    ret.msg = 'Disculpe, pero no se determinó la acción a realizar.';
+    statusCode = 422;
+  }
+  else if (code === 'slug') {
+    ret.msg = 'Disculpe, pero el curso seleccionado es incorrecto.';
+    statusCode = 422;
+  }
+  else if (code === 'wasNotPreviousCourse') {
+    ret.msg = `Disculpe, pero no puede visualizar el contenido. Debe finalizar los cursos previos a este.`;
+    statusCode = 422
+  }
+  else if (code === 'wasRealized') {
+    ret.msg = `Disculpe, pero ya has realizado esta acción anteriormente.`;
+    statusCode = 422;
+  }
+  else if (code === 'wasRealizedAllTest') {
+    ret.msg = `Disculpe, pero ya ha aprobado todos los exámenes de este curso.`;
+    statusCode = 422;
+  }
+  else if (code === 'wasRealizedTest') {
+    ret.msg = `Disculpe, pero ya ha aprobado este examen anteriormente.`;
+    statusCode = 422;
+  }
+
+  return res.status(statusCode).json(ret);
+}
+
+function checkAnswerCheckbox(value: string | null, correctAnswer: number | null): boolean {
+  if (value && value.indexOf(',') > -1) {
+    const ans = value.split(',') || []; // separate values
+    if (correctAnswer !== null) {
+      // check if assignate points if the correctAnswer exists in the received answer
+      return ans && ans.indexOf(correctAnswer.toString()) > -1
+    }
+  }
+  return false;
+}
+
+export function setPointToTest(question: any, inputAnswer: any): number {
+  if (question.inputType === 'checkbox') {
+    // check if answer cotains comma (,)
+    if (inputAnswer.answer && inputAnswer.answer.indexOf(',') > -1) {
+      // assignate points if the correctAnswer exists in the received answer
+      return checkAnswerCheckbox(inputAnswer.answer, question.correctAnswer) ? 1 : 0;
+    }
+  }
+
+  // check if questions has a default answer
+  if (question.correctAnswer !== null)
+    return question.correctAnswer.toString() === inputAnswer.answer ? 1 : 0;
+
+  return 1;
 }
